rewrite GUI
