+++ phishpedia_gui.py
@@ -1,173 +1,9 @@
 import sys
-from PyQt5.QtWidgets import QApplication, QWidget, QVBoxLayout, QLabel, QLineEdit, QPushButton, QFileDialog, QTextEdit, QHBoxLayout
-from PyQt5.QtGui import QPixmap, QImage, QIcon, QFont
-from PyQt5.QtCore import Qt
-from phishpedia import PhishpediaWrapper
-import cv2
-from PIL import Image
-import numpy as np
-
-class PhishpediaGUI(QWidget):
-    def __init__(self):
-        super().__init__()
-        self.initUI()
-        self.phishpedia_cls = PhishpediaWrapper()
-        self.current_pixmap = None
-
-    def initUI(self):
-        self.setWindowTitle('Phishpedia GUI')
-        self.setGeometry(100, 100, 800, 600)
-
-        main_layout = QVBoxLayout()
-
-        # URL Input
-        url_layout = QHBoxLayout()
-        self.url_label = QLabel('Enter URL:')
-        url_layout.addWidget(self.url_label)
-        self.url_input = QLineEdit()
-        url_layout.addWidget(self.url_input)
-        main_layout.addLayout(url_layout)
-
-        # Image Upload
-        image_layout = QHBoxLayout()
-        self.image_label = QLabel('Upload Screenshot:')
-        image_layout.addWidget(self.image_label)
-        self.image_input = QLineEdit()
-        image_layout.addWidget(self.image_input)
-        self.image_button = QPushButton('Browse')
-        self.image_button.clicked.connect(self.upload_image)
-        image_layout.addWidget(self.image_button)
-        main_layout.addLayout(image_layout)
-
-        # Detect Button
-        self.detect_button = QPushButton('Detect')
-        self.detect_button.clicked.connect(self.detect_phishing)
-        main_layout.addWidget(self.detect_button)
-
-        # Result Display
-        result_layout = QHBoxLayout()
-        self.result_label = QLabel('Detection Result:')
-        result_layout.addWidget(self.result_label)
-        self.result_display = QTextEdit()
-        self.result_display.setReadOnly(True)
-        self.result_display.setFixedHeight(150)  # 设置固定高度
-        result_layout.addWidget(self.result_display)
-        main_layout.addLayout(result_layout)
-
-        # Visualization Display
-        visualization_layout = QVBoxLayout()
-        self.visualization_label = QLabel('Visualization Result:')
-        self.visualization_label.setFixedHeight(self.visualization_label.fontMetrics().height())  # 设置固定高度为一行字的高度
-        visualization_layout.addWidget(self.visualization_label)
-        self.visualization_display = QLabel()
-        self.visualization_display.setAlignment(Qt.AlignCenter)
-        self.visualization_display.setMinimumSize(600, 400)  # 设置最小大小
-        visualization_layout.addWidget(self.visualization_display)
-        main_layout.addLayout(visualization_layout)
-
-        self.setLayout(main_layout)
-
-        # Apply stylesheet
-        self.setStyleSheet("""
-            QLabel {
-                color: #333;
-            }
-            QLineEdit, QTextEdit {
-                border: 1px solid #ccc;
-                padding: 5px;
-                border-radius: 4px;
-            }
-            QPushButton {
-                background-color: #007BFF;
-                color: white;
-                border: none;
-                padding: 8px 16px;
-                border-radius: 4px;
-            }
-            QPushButton:hover {
-                background-color: #0056b3;
-            }
-        """)
-
-        # Connect resize event to handle image scaling
-        self.resizeEvent = self.on_resize
-
-        # Set dynamic font size based on screen DPI
-        self.set_dynamic_font_size()
-
-    def set_dynamic_font_size(self):
-        screen = QApplication.primaryScreen()
-        dpi = screen.logicalDotsPerInch()
-        base_font_size = 18  # Base font size for 96 DPI
-        font_size = base_font_size * (dpi / 150)
-
-        font = QFont()
-        font.setPointSizeF(font_size)
-
-        for widget in self.findChildren(QLabel) + self.findChildren(QLineEdit) + self.findChildren(QPushButton) + [self.result_display]:
-            widget.setFont(font)
-
-    def upload_image(self):
-        options = QFileDialog.Options()
-        file_name, _ = QFileDialog.getOpenFileName(self, "Select Screenshot", "", "Images (*.png *.jpg *.jpeg)", options=options)
-        if file_name:
-            self.image_input.setText(file_name)
-
-    def detect_phishing(self):
-        url = self.url_input.text()
-        screenshot_path = self.image_input.text()
-
-        if not url or not screenshot_path:
-            self.result_display.setText("Please enter URL and upload a screenshot.")
-            return
-
-        phish_category, pred_target, matched_domain, plotvis, siamese_conf, pred_boxes, logo_recog_time, logo_match_time = self.phishpedia_cls.test_orig_phishpedia(url, screenshot_path, None)
-
-        # 根据 phish_category 改变颜色
-        phish_category_color = 'green' if phish_category == 0 else 'red'
-        result_text = f'<span style="color: {phish_category_color};">Phish Category(0 for benign, 1 for phish, default is benign): {phish_category}</span><br>'
-        result_text += f"Predicted Target: {pred_target}<br>"
-        result_text += f"Matched Domain: {matched_domain}<br>"
-        result_text += f"Siamese Confidence: {siamese_conf}<br>"
-        result_text += f"Logo Recognition Time: {logo_recog_time} seconds<br>"
-        result_text += f"Logo Match Time: {logo_match_time} seconds<br>"
-
-        self.result_display.setText(result_text)
-
-        if phish_category == 1 and plotvis is not None:
-            self.display_image(plotvis)
-        if phish_category == 0:
-            # 展示已经上传的屏幕截图
-            self.display_image(plotvis)
-
-    def display_image(self, plotvis):
-        try:
-            # Convert plotvis to QImage
-            height, width, channel = plotvis.shape
-            bytes_per_line = 3 * width
-            plotvis_qimage = QImage(plotvis.data, width, height, bytes_per_line, QImage.Format_RGB888)
-            
-            # Convert QImage to QPixmap
-            self.current_pixmap = QPixmap.fromImage(plotvis_qimage)
-            self.update_image_display()
-        except Exception as e:
-            print(f"Error converting image: {e}")
-
-    def update_image_display(self):
-        if self.current_pixmap:
-            # Calculate the available space for the image
-            available_width = self.width()
-            available_height = self.height() - self.visualization_label.geometry().bottom() - 50
-
-            # Calculate the scaled size while maintaining aspect ratio
-            scaled_pixmap = self.current_pixmap.scaled(available_width, available_height, Qt.KeepAspectRatio, Qt.SmoothTransformation)
-            self.visualization_display.setPixmap(scaled_pixmap)
-
-    def on_resize(self, event):
-        self.update_image_display()
+from PyQt5.QtWidgets import QApplication
+from GUI.ui import PhishpediaUI
 
 if __name__ == '__main__':
     app = QApplication(sys.argv)
-    ex = PhishpediaGUI()
+    ex = PhishpediaUI()
     ex.show()
     sys.exit(app.exec_())
